Changed percentage display to an int

+++ lpbot/modules/reddit.py
@@ -66,7 +66,7 @@ def rpost_info(bot, trigger, match=None):
     else:
         point_color = colors.RED
 
-    percent = color(unicode(s.upvote_ratio * 100) + '%', point_color)
+    percent = color(unicode(int(s.upvote_ratio * 100)) + '%', point_color)
 
     message = message.format(
         title=s.title, link=link, nsfw=nsfw, points=s.score, percent=percent,
